marker maps in shopFragment and login with email and password function from firebase authentication

+++ Passtick3/app/src/main/java/coe/project/passtick/Shops.kt
@@ -6,8 +6,8 @@ import com.google.firebase.database.IgnoreExtraProperties
 data class Shops (
     val cost: Long?=0,
     val donate: Long?=0,
-    val lat: Long?=0,
-    val lng: Long?=0,
+    val lat: Double?=0.0,
+    val lng: Double?=0.0,
     val logo: String?="",
     val outstanding: Long?=0,
     val pieces: Long?=0,
